Added new script for counting residues in PDB file
This Python script processes Protein Data Bank (PDB) files to identify and count the number of specific and total residues. It provides a count of water residues (labeled as 'HOH' or 'WAT'), lipid residues (labeled as 'PA', 'PC', or 'OL'), and all other residues separately. Additionally, the script identifies the range of lipid residues in the protein structure. It uses a simple line-by-line parsing method to go through each 'ATOM' line in the PDB file, extracting the residue name and number. The script is designed to be run from the command line, taking the path to the PDB file as an argument.